Add a fqdn to etc/hosts, speeds up sendmail
Sendmail will wait 60 seconds if its hostname is not a fqdn and it can't
find an alias in /etc/hosts that is a fqdn. This slows down running the
exploit script.

Fix this by changing /etc/hosts in the entrypoint to add myhost.domain
to the first non-127.0.0.1 ipv4 ip line in /etc/hosts.